Rename crypt.ElGamalEngine to crypt.ElGamalEngineK in order not to confuse with the bouncycastle original file
