--- issues/ElGamalEngineK-is-never-used/Description
@@ -1,1 +1,0 @@
-crypt.utils.ElGamalEngineK is never used. Should be removed?
